Update oe2pg.py
Updated DBMirror to behave as an incremental update utility as opposed to a "as it is now" tool. 

The DBMirror class now parses both the original and new table looking for differences. If it finds that there are data not present in the new table, it mirrors it.

This update effectively makes this an update tool that can run indefinitely to create a copy of an OE database on PostgreSQL.


TO-DO: Further revise the data types. I believe there are a couple OE conventions I haven't fixed yet.